[level 2] Title: n^2 배열 자르기, Time: 6.33 ms, Memory: 97 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -4,8 +4,8 @@ class Solution {
         int[] ans = new int[size];
         int idx = 0;
         for (long i = left; i <= right; i++) {
-            long row = i / n + 1;
-            long col = i % n + 1;
+            int row = (int)(i / n) + 1;
+            int col = (int)(i % n) + 1;
             //System.out.printf("at i: %d, row:%d, col:%d\n", i, row, col);
             
             ans[idx] = row < col ? (int)col : (int)row;
